+++ src/qnethernet/QNEthernetUDP.cpp
@@ -385,8 +385,9 @@ int EthernetUDP::beginPacket(const IPAddress ip, const uint16_t port) {
 int EthernetUDP::beginPacket(const char* const host, const uint16_t port) {
 #if LWIP_DNS
   IPAddress ip;
-  if (!DNSClient::getHostByName(host, ip,
-                                QNETHERNET_DEFAULT_DNS_LOOKUP_TIMEOUT)) {
+  if (!DNSClient::getHostByName(
+          host, ip,
+          static_cast<uint32_t>(QNETHERNET_DEFAULT_DNS_LOOKUP_TIMEOUT))) {
     return false;
   }
   return beginPacket(ip, port);
@@ -488,8 +489,9 @@ bool EthernetUDP::send(const char* const host, const uint16_t port,
                        const void* const data, const size_t len) {
 #if LWIP_DNS
   IPAddress ip;
-  if (!DNSClient::getHostByName(host, ip,
-                                QNETHERNET_DEFAULT_DNS_LOOKUP_TIMEOUT)) {
+  if (!DNSClient::getHostByName(
+          host, ip,
+          static_cast<uint32_t>(QNETHERNET_DEFAULT_DNS_LOOKUP_TIMEOUT))) {
     return false;
   }
   return send(ip, port, data, len);
